Web tools
Moved loadS3 and loadHTTP to new file, web_tools.py. MaskGenUI now
imports this.
